pyproject.toml
[tool.poetry]
authors = ["Jean-Christophe Bianic <jc.bianic@gmail.com>"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Web Environment",
  "Framework :: Flask",
  "Framework :: Pydantic",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Software Development",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Internet",
  "Typing :: Typed",
]
description = "A Flask extension, inspired by FastAPI that uses Pydantic to provide easy-to-configure data validation for request parsing and response serialization."
documentation = "https://flask-jeroboam.readthedocs.io"
homepage = "https://github.com/jcbianic/flask-jeroboam"
license = "MIT"
name = "flask-jeroboam"
readme = "README.md"
repository = "https://github.com/jcbianic/flask-jeroboam"
version = "0.1.0b3"

[tool.poetry.urls]
Changelog = "https://github.com/jcbianic/flask-jeroboam/releases"

[tool.poetry.dependencies]
Flask = "^3.1.0"
pydantic1 = "^1.10.3"
python = "^3.9.1"
wheel = "^0.45.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.workflows]
optional = true

[tool.poetry.group.dev.dependencies]
GitPython = "^3.1.30"
Pallets-Sphinx-Themes = "^2.0.2"
Pygments = ">=2.10.0"
black = ">=24.10.0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = "^6.0.0"
flake8-bandit = "4.1.1"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
isort = ">=5.10.1"
mypy = ">=0.930"
myst-parser = {version = ">=0.16.1"}
nox = "^2024.10.9"
nox-poetry = "^1.0.3"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pycln = "^2.1.1"
pytest = ">=6.2.5"
pytest-mock = "^3.8.2"
pyupgrade = ">=2.29.1"
safety = "^3.2.14"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
sphinx-multiversion = "^0.2.4"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
certifi = "2024.12.14"
sourcery-cli = "^1.20.0"
flake8-pyi = "^23.1.2"
python-dotenv = "^0.21.1"
pydantic1 = {extras = ["email-validator"], version = "^1.10.5"}
email-validator = "^1.3.1"
types-toml = "^0.10.8.4"
toml = "^0.10.2"

[tool.poetry.group.docs.dependencies]
sphinx = "<5.3.0"
sphinx-click = "^4.4.0"
myst_parser = "^0.18.1"
pallets_sphinx_themes= "^2.0.2"
sphinx_multiversion = "^0.2.4"
sphinx-tabs = "^3.4.1"
sphinx-copybutton = "^0.5.1"

[tool.poetry.group.workflows.dependencies]
pip = "^24.3.1"
nox = "^2024.10.9"
nox-poetry = "^1.0.3"
poetry = "^2.0"
virtualenv = "20.29.0"

[tool.coverage.paths]
source = ["flask_jeroboam", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
relative_files = true
source = ["flask_jeroboam"]

[tool.coverage.report]
fail_under = 100
show_missing = true

[tool.isort]
force_single_line = true
lines_after_imports = 2
profile = "black"

[tool.mypy]
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_unreachable = true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.2.2"]


[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
